Django-Wysiwyg/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
distlib==0.3.1
Django==2.2.16
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-dotenv==1.4.2
django-extensions==2.2.9
django-froala-editor==3.2.2
django-js-asset==1.2.2
django-pagedown==2.2.0
django-quill-editor==0.1.10
django-storages==1.10.1
django-summernote==0.8.11.6
django-tinymce==3.0.2
djangorestframework==3.11.1
filelock==3.0.12
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pipenv==2020.8.13
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
virtualenv==20.0.34
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.3.0
